pyproject.toml
[project]
name = "webtapuapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "camelot-py>=1.0.9",
    "flask>=3.1.2",
    "lxml>=6.0.2",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.7",
    "pyicu>=2.15.3",
    "pymupdf>=1.26.5",
    "regex>=2025.9.18",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "unidecode>=1.4.0",
]
